scanner/requirements.txt
flask==2.1.0
numpy==1.21.0
joblib==1.0.1
scikit-learn==0.24.2
python-magic==0.4.24
flask-cors==3.0.10
python-dotenv==0.19.0
requests==2.26.0
pefile==2021.5.24
pandas==1.3.0
werkzeug==2.1.0
